+++ api/src/main/java/vook/server/api/web/routes/user/UserApi.java
@@ -13,6 +13,7 @@ import vook.server.api.web.common.CommonApiResponse;
 import vook.server.api.web.routes.user.reqres.UserInfoResponse;
 import vook.server.api.web.routes.user.reqres.UserOnboardingRequest;
 import vook.server.api.web.routes.user.reqres.UserRegisterRequest;
+import vook.server.api.web.routes.user.reqres.UserUpdateInfoRequest;
 import vook.server.api.web.swagger.ComponentRefConsts;
 
 @Tag(name = "user", description = "사용자 관련 API")
